libbinder: Java/C++ *_TRANSACTION parity
TWEET_TRANSACTION and LIKE_TRANSACTION have long been Java-only features
in Android. Though, by all accounts, it has been 9.5 years since their
inclusion, and the social landscape has been made portrait (and back and
forth, and back and forth, and so goes the pendelum of time), the
managed app, now (!) every binder and daemon, every process, thread,
every fiber, every bound iface, whether preinstalled or downloaded,
enjoys itself a quiet smile or a chuckle or a warm feeling in its heart
at the directed tweet (or cold rage or fury when the need to argue
arises!). Same too, with the ready-fire like, an uncrafted promotor of
unconstrained joy and appreciation for the IPCs or APIs, carefully
crafted and readily available, or the interesting return value, or heck,
maybe even the timing or just for the lols!

Though, due to timing constraints (and the NDK/Rust backend
abstraction from transaction codes in favor of function APIs which
encode the protocol of these functions), Android S might not see this
functionality there, it is a future consideration.

Fixes: 178706440
Test: N/A
Change-Id: I3f628c747b70a77e199abc7252a56fdc36ef4308
Signed-off-by: Akash Srivastava <akashniki@gmail.com>